+++ src/pkgdevx/logging/__init__.py
@@ -1,5 +1,5 @@
 """__init__ for pkgdevx logging module."""
 
-from ._logging import configure_logging
+from ._logging import GLOBAL_CONSOLE, configure_logging
 
-__all__: list[str] = ["configure_logging"]
+__all__: list[str] = ["GLOBAL_CONSOLE", "configure_logging"]